create comment modal
